Update imports after scanAllGuides move and delete legacy utils.ts
