+++ guides/sync-use-cases.ts
@@ -6,8 +6,7 @@ import matter from 'gray-matter';
 import { fileURLToPath } from 'url';
 import { validateMacros } from '../serving/mcp-server/lib/macros.ts';
 import { validateFeature } from '../serving/mcp-server/data/baseline.ts';
-import { scanAllGuides, type GuideInventory } from '../harness/lib/utils.ts';
-import { ProjectStatus, type PreparedGuide, type GuideInventoryResult, validateGuide, processGuideInventory, getStatusName } from '../lib/guide-validation.ts';
+import { ProjectStatus, type PreparedGuide, type GuideInventoryResult, validateGuide, processGuideInventory, getStatusName, scanAllGuides, type GuideInventory } from '../lib/guide-validation.ts';
 
 // --- Types ---
