+++ guides/validation.test.ts
@@ -3,8 +3,7 @@ import assert from 'node:assert';
 import path from 'node:path';
 
 // Import shared utilities
-import { scanAllGuides } from '../harness/lib/utils.ts';
-import { processGuideInventory } from '../lib/guide-validation.ts';
+import { scanAllGuides, processGuideInventory } from '../lib/guide-validation.ts';
 
 const REPO_ROOT = path.resolve(import.meta.dirname, '..');
