+++ package.json
@@ -10,6 +10,7 @@
   },
   "scripts": {
     "test": "pnpm -r test",
+    "install:gray-matter": "pnpm add -D -w gray-matter",
     "baselinestatus": "pnpm --filter modern-web-mcp baselinestatus",
     "lint": "oxlint --ignore-path .oxlintignore",
     "typecheck": "pnpm -r typecheck",
@@ -59,6 +60,7 @@
     "@types/node": "^25.5.0",
     "@types/omelette": "^0.4.5",
     "gh-pages": "^6.3.0",
+    "gray-matter": "^4.0.3",
     "oxlint": "^1.55.0",
     "typescript": "^5.9.3"
   },
