+++ serving/benchmarks/rag/generate-eval-queries.ts
@@ -2,7 +2,7 @@ import fs from "fs";
 import path from "path";
 import matter from "gray-matter";
 import { fileURLToPath } from "url";
-import { classifyGuide, scanAllGuides } from "../../harness/lib/utils.ts";
+import { classifyGuide, scanAllGuides } from "../../../lib/guide-validation.ts";
 
 const __filename = fileURLToPath(import.meta.url);
 const __dirname = path.dirname(__filename);
