+++ serving/skills-cli/build-dist.ts
@@ -2,7 +2,7 @@ import fs from "fs";
 import path from "path";
 import { execSync } from "child_process";
 import matter from "gray-matter";
-import { classifyGuide, scanAllGuides } from "../../harness/lib/utils.ts";
+import { classifyGuide, scanAllGuides } from "../../lib/guide-validation.ts";
 import { getFeatureName } from "../mcp-server/data/baseline.ts";
 
 const ROOT_DIR = path.resolve(import.meta.dirname, "../.."); // guidance/
